Removed keys and restructured the code
Removed keys and restructured the js-client code also added instructions for deploying in the readme file
