+++ .env
@@ -8,11 +8,11 @@ NODE_ADDRESS=http://159.65.118.250:7777/rpc
 EVENT_STREAM_ADDRESS=http://159.65.118.250:9999/events/main
 RECEIVER_ACCOUNT_ONE=017e82abcc9539a01cfd9d63ae8c9c8b3a752a6f75ba1ab148714eea03e9be69a7
 
-LIQUIDITYTRANSFORMER_WASM_PATH=JsClients/LIQUIDITYTRANSFORMER/wasm/liquidity_transformer.wasm
-WISETOKEN_WASM_PATH=JsClients/WISETOKEN/wasm/stakeabletoken.wasm
+LIQUIDITYTRANSFORMER_WASM_PATH=Scripts/LIQUIDITYTRANSFORMER/wasm/liquidity_transformer.wasm
+WISETOKEN_WASM_PATH=Scripts/WISETOKEN/wasm/stakeabletoken.wasm
 
-LIQUIDITYTRANSFORMER_MASTER_KEY_PAIR_PATH=JsClients/LIQUIDITYTRANSFORMER/keys/
-WISETOKEN_MASTER_KEY_PAIR_PATH=JsClients/WISETOKEN/keys/
+LIQUIDITYTRANSFORMER_MASTER_KEY_PAIR_PATH=Scripts/LIQUIDITYTRANSFORMER/keys/
+WISETOKEN_MASTER_KEY_PAIR_PATH=Scripts/WISETOKEN/keys/
 
 LIQUIDITYTRANSFORMER_CONTRACT_HASH=177d86764f06936ef8af744858cf1f6b1fd88f77b86a9f47f53e9d6b03499385
 LIQUIDITYTRANSFORMER_PACKAGE_HASH=84f06cf0dd25c02109e460ab9330cb28675cfce60a0936dfec9390886b19fabe
