--- JsClients/LIQUIDITYTRANSFORMER/keys/public_key.pem
@@ -1,3 +1,0 @@
------BEGIN PUBLIC KEY-----
-MCowBQYDK2VwAyEAuTSihx6sCpsEFph+8EF1HWMyhGOcwjJNbSwFfH4h1Nw=
------END PUBLIC KEY-----
