+++ JsClients/LIQUIDITYTRANSFORMER/src/types.d.ts
@@ -4,5 +4,5 @@ export type RecipientType = CLPublicKey | CLAccountHash | CLByteArray;
 
 export interface IPendingDeploy {
   deployHash: string;
-  deployType: LIQUIDITYEvents;
+ // deployType: LIQUIDITYEvents;
 }
