--- JsClients/WISETOKEN/keys/public_key_hex
@@ -1,1 +1,0 @@
-01b934a2871eac0a9b0416987ef041751d633284639cc2324d6d2c057c7e21d4dc
