--- JsClients/WISETOKEN/keys/secret_key.pem
@@ -1,3 +1,0 @@
------BEGIN PRIVATE KEY-----
-MC4CAQAwBQYDK2VwBCIEICie3PaBJXdTM6w0eWtvdGrWMTRnLi1Wo6a80hshJHGT
------END PRIVATE KEY-----
