+++ Scripts/LIQUIDITYTRANSFORMER/deploy/liquidityTransformerContract.ts
@@ -1,6 +1,6 @@
 import { config } from "dotenv";
 config();
-import { WISETokenClient, utils, constants } from "../src";
+import { LIQUIDITYClient, utils, constants } from "../../../JsClients/LIQUIDITYTRANSFORMER/src";
 import { parseTokenMeta, sleep, getDeploy } from "./utils";
 
 import { Keys } from "casper-js-sdk";
@@ -9,43 +9,41 @@ const {
 	NODE_ADDRESS,
 	EVENT_STREAM_ADDRESS,
 	CHAIN_NAME,
-	WISETOKEN_WASM_PATH,
-	WISETOKEN_MASTER_KEY_PAIR_PATH,
-	WISETOKEN_INSTALL_PAYMENT_AMOUNT,
-	WISETOKEN_CONTRACT_NAME,
+	LIQUIDITYTRANSFORMER_WASM_PATH,
+	LIQUIDITYTRANSFORMER_MASTER_KEY_PAIR_PATH,
+	LIQUIDITYTRANSFORMER_INSTALL_PAYMENT_AMOUNT,
+	LIQUIDITYTRANSFORMER_CONTRACT_NAME,
 	WCSPR_ADDRESS,
 	SYNTHETIC_CSPR_ADDRESS,
 	PAIR_ADDRESS,
 	ROUTER_ADDRESS,
-	FACTORY_ADDRESS,
-	LIQUIDITY_GUARD_ADDRESS,
-	LAUNCH_TIME
+	ROUTER_PACKAGE_HASH,
+	WISETOKEN_CONTRACT_HASH
 } = process.env;
 
 const KEYS = Keys.Ed25519.parseKeyFiles(
-	`${WISETOKEN_MASTER_KEY_PAIR_PATH}/public_key.pem`,
-	`${WISETOKEN_MASTER_KEY_PAIR_PATH}/secret_key.pem`
+	`${LIQUIDITYTRANSFORMER_MASTER_KEY_PAIR_PATH}/public_key.pem`,
+	`${LIQUIDITYTRANSFORMER_MASTER_KEY_PAIR_PATH}/secret_key.pem`
 );
 
 const test = async () => {
-	const wise = new WISETokenClient(
+	const liquidity = new LIQUIDITYClient(
 		NODE_ADDRESS!,
 		CHAIN_NAME!,
 		EVENT_STREAM_ADDRESS!
 	);
-	
-	const installDeployHash = await wise.install(
+
+	const installDeployHash = await liquidity.install(
 		KEYS,
 		WCSPR_ADDRESS!,
 		SYNTHETIC_CSPR_ADDRESS!,
 		PAIR_ADDRESS!,
 		ROUTER_ADDRESS!,
-		FACTORY_ADDRESS!,
-		LIQUIDITY_GUARD_ADDRESS!,
-		LAUNCH_TIME!,
-		WISETOKEN_CONTRACT_NAME!,
-		WISETOKEN_INSTALL_PAYMENT_AMOUNT!,
-		WISETOKEN_WASM_PATH!
+		ROUTER_PACKAGE_HASH!,
+		WISETOKEN_CONTRACT_HASH!,
+		LIQUIDITYTRANSFORMER_CONTRACT_NAME!,
+		LIQUIDITYTRANSFORMER_INSTALL_PAYMENT_AMOUNT!,
+		LIQUIDITYTRANSFORMER_WASM_PATH!
 	);
 
 	console.log(`... Contract installation deployHash: ${installDeployHash}`);
@@ -61,14 +59,14 @@ const test = async () => {
 
 	const contractHash = await utils.getAccountNamedKeyValue(
 		accountInfo,
-		`${WISETOKEN_CONTRACT_NAME!}_contract_hash`
+		`${LIQUIDITYTRANSFORMER_CONTRACT_NAME!}_contract_hash`
 	);
 
 	console.log(`... Contract Hash: ${contractHash}`);
 
 	const packageHash = await utils.getAccountNamedKeyValue(
 		accountInfo,
-		`${WISETOKEN_CONTRACT_NAME!}_package_hash`
+		`${LIQUIDITYTRANSFORMER_CONTRACT_NAME!}_package_hash`
 	);
 	
 	console.log(`... Package Hash: ${packageHash}`);
