+++ Scripts/WISETOKEN/deploy/wiseContractFunction.ts
@@ -1,6 +1,6 @@
 import { config } from "dotenv";
 config();
-import { WISETokenClient, utils, constants } from "../src";
+import { WISETokenClient, utils, constants } from "../../../JsClients/WISETOKEN/src";
 import { parseTokenMeta, sleep, getDeploy } from "./utils";
 
 import {
@@ -238,32 +238,3 @@ const test = async () => {
 };
 
 //test();
-
-export const balanceOf = async (contractHash:string, key:string) => {
-  
-	console.log(`... Contract Hash: ${contractHash}`);
-  
-	// We don't need hash- prefix so i'm removing it
-	await wise.setContractHash(contractHash);
-  
-   //balanceof
-	let balance = await wise.balanceOf(key);
-  
-	console.log(`... Balance: ${balance}`);
-  
-	return balance;
-  
-  };
-
-export const getTotalSupply = async (contractHash:string) => {
-  
-	// We don't need hash- prefix so i'm removing it
-	await wise.setContractHash(contractHash);
-  
-	 //totalsupply
-	 let totalSupply = await wise.totalSupply();
-	 console.log(contractHash +` = ... Total supply: ${totalSupply}`);
-  
-	return totalSupply;
-	
-};
